Ajustes para evitar conflitos de merge e economizar linhas de codigo
